doc: move instructions to its own directory.
